% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyToPDB.R
\name{applyToPDB}
\alias{applyToPDB}
\title{Applies a function over a list of PDB ID entries.}
\usage{
applyToPDB(
  listpdb = NULL,
  FUNCTION,
  as.df = TRUE,
  cores = 1,
  progressbar = TRUE,
  ...
)
}
\arguments{
\item{listpdb}{A list/vector containing the PDB IDs of interest. If NULL,
the complete list of PDB entries is downloaded and used.}

\item{FUNCTION}{A function of interest.}

\item{as.df}{A logical that stands for "as.data.frame". If TRUE, the output
will be returned in the form of a data.frame, otherwise a list.}

\item{cores}{Number of CPU cores to be used.}

\item{progressbar}{A logical to print in screen a progress bar.}

\item{...}{optional arguments to FUNCTION.}
}
\value{
A data.frame with the PDB IDs (first colunm) and the output of the
function of interest (second column) or a list with the results.
}
\description{
Given a function of interest, it is applied to all the PDB entries.
See supported functions in ?queryFunctions.
}
\examples{
    listpdb <- c("1s72", "1bau", "1rna")
    applyToPDB(listpdb, queryTechnique)

}
\author{
Diego Gallego
}
