% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNuc.R
\name{checkNuc}
\alias{checkNuc}
\title{Check nucleotides}
\usage{
checkNuc(
  pdb,
  model = 1,
  chain = "all",
  id = NULL,
  refatm = "C4'",
  force = FALSE
)
}
\arguments{
\item{pdb}{A pdb object as obtained from cifAsPDB or read.cif/read.pdb
(bio3d package).}

\item{model}{A string with the desired model number.}

\item{chain}{A string with the desired chain id.}

\item{id}{A string with the ID of the input pdb structure.}

\item{refatm}{A string with the atom to use to identify the nucelotides.
Important to analyse models with just (in example) phosphate atoms,
in which refatm should be set to "P" (it was thought when analysing
the structure with PDB code: 1Y1Y).}

\item{force}{A logical to force the analysis. Useful when the function
does not recognise a nucleic acid in the structure (e.g. because all
bases are non-canonical: 1PBL, 1XV6, 1DV4, etc).}
}
\value{
A data.frame with the data for every nucleotide.
}
\description{
From a nucleic acid structure (pdb object), it checks the presence of all
nucleotide atoms, bond distances, chain breaks and others.
}
\examples{
    data <- checkNuc(cifAsPDB("1am0"))
    broken <- which(data$Break == TRUE)
    data[broken, ] ## See the places in which the chain is broken

}
\author{
Diego Gallego
}
