% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanByPucker.R
\name{cleanByPucker}
\alias{cleanByPucker}
\title{Subset nucleotide data according with puckering}
\usage{
cleanByPucker(
  ntinfo,
  surenorth = FALSE,
  suresouth = FALSE,
  pucker = "C3'endo",
  range = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ntinfo}{A data.frame. The output of
\code{\link[=pipeNucData]{pipeNucData()}}.}

\item{surenorth}{A logical to return nucleotides in north with restrictions
in delta and Dp.}

\item{suresouth}{A logical to return nucleotides in south with restrictions
in delta and Dp.}

\item{pucker}{A string with the puckering state of interest. Only necessary
if surenorth and suresouth are FALSE and range is NULL. When using this
option, only the phase is used to subset the data.}

\item{range}{A numeric vector with the desired phaserange. Only used if no
other argument above could be applyed.}

\item{verbose}{A logical to print details of the process.}
}
\value{
An integer vector with the nucleotides (ntID) matching the desired
puckering state.
}
\description{
Function to clean raw data after the pipeline
\code{\link[=pipeNucData]{pipeNucData()}}.
It takes a data.frame that should have the columns "pu_phase", "delta" and
"Dp", and returns the nucleotides matching the desired puckering state.
}
\examples{
    ntinfo <- pipeNucData("1bau")
    north_ntID <- cleanByPucker(ntinfo, surenorth=TRUE)
    north <- ntinfo[ntinfo$ntID \%in\% north_ntID,]

}
\author{
Diego Gallego
}
