% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dataset_wadley2007.R
\docType{data}
\name{dataset_wadley2007}
\alias{dataset_wadley2007}
\title{Dataset of RNA structures by Wadley et al.}
\format{An object of class data.frame
\describe{
\item{pdb:}{PDB ID.}
\item{chain:}{Chain}
}}
\usage{
data(dataset_wadley2007)
}
\value{
data.frame with list of PDB IDs.
}
\description{
Data.frame containing the list of PDB IDs of the structures used by
Wadley et al. in their eta-theta study about RNA.
}
\references{
\preformatted{Wadley, L.M., K.S. Keating, C.M Duarte, and A.M. Pyle. 2007.
“Evaluating and Learning from Rna Pseudotorsional Space: 
Quantitative Validation of a Reduced Representation for
RNA Structure.” Journal of Molecular Biology 372 (4): 942–57
}
}
\keyword{datasets}
