% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.extract_ntinfo}
\alias{.extract_ntinfo}
\title{Obtain nucleotide details from a data set of RNA structures}
\usage{
.extract_ntinfo(pdbID, ...)
}
\arguments{
\item{pdbID}{A list/vector containing the desired PDB IDs or a list of pdb
objects as provided by "read.pdb", "read.cif", "cifParser" ...}

\item{...}{Arguments to be passed to the checkNuc function}
}
\value{
A data.frame with data about every nucleotide in the input set.
}
\description{
Pipeline to generate a data.frame with the desired info for a list of PDB.
Nucleotides are labeled with a unique identifier (column ntID).
}
\examples{
    pdblist <- list("1bau", "2rn1")
    ntinfo <- extract_ntinfo(pdbID=pdblist)

}
\author{
Eric Matamoros & Diego Gallego
}
