% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CIF.R
\name{eRMSD}
\alias{eRMSD}
\alias{eRMSD,CIF,CIF-method}
\alias{eRMSD,ANY,ANY-method}
\title{Compute the epsilon RMSD between two RNA structures}
\usage{
eRMSD(
  cif1 = NULL,
  cif2 = NULL,
  pdb1 = NULL,
  pdb2 = NULL,
  rvectors1 = NULL,
  rvectors2 = NULL
)

\S4method{eRMSD}{CIF,CIF}(cif1 = NULL, cif2 = NULL)

\S4method{eRMSD}{ANY,ANY}(pdb1 = NULL, pdb2 = NULL, rvectors1 = NULL, rvectors2 = NULL)
}
\arguments{
\item{cif1}{A CIF object as otained from
\code{\link[=cifParser]{cifParser()}}.}

\item{cif2}{A CIF object as otained from
\code{\link[=cifParser]{cifParser()}}.}

\item{pdb1}{A pdb object as obtained from cifAsPDB or read.cif/read.pdb
(from bio3d package).}

\item{pdb2}{A pdb object as obtained from cifAsPDB or read.cif/read.pdb
(from bio3d package).}

\item{rvectors1}{A data.frame as obtained from
\code{\link[=rVector]{rVector()}}
using simple_out=TRUE.}

\item{rvectors2}{A data.frame as obtained from
\code{\link[=rVector]{rVector()}}
using simple_out=TRUE.}
}
\value{
A numeric with the epsilon RMSD between the two structures.
}
\description{
Given two RNA with the same length, the functions calculates its epsilon
RMSD (Bottaro et al. 2014), reproducing BaRNAba software (Bottaro et al.
2018).
}
\examples{
    cif <- cifParser("2d18")
    model1 <- selectModel(cif=cif, model=1)
    model3 <- selectModel(cif=cif, model=3)
    eRMSD(cif1=model1, cif2=model3)

}
\references{
\preformatted{Bottaro, S., F. Di Palma, and G Bussi. 2014. “The Role of Nucleobase 
Interactions in RNA Structure and Dynamics.” Nucleic Acids Research 
42 (21): 13306–14
Bottaro, S., G. Bussi, G. Pinamonti, S. Reiber, W. Boomsma, and K. 
Lindorff-Larsen. 2018. “Barnaba: Software for Analysis of Nucleic 
Acids Structures and Trajectories.” RNA, Epub ahead of print
}
}
\author{
Diego Gallego
}
