% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBindingSite.R
\name{findBindingSite}
\alias{findBindingSite}
\title{Function to get data about the atoms in interacting site.}
\usage{
findBindingSite(
  pdb,
  cutoff = 5,
  select = "Nuc",
  nchain = NULL,
  pchain = NULL,
  hydrogens = FALSE,
  byres = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{pdb}{A cif/pdb object obtained from cifParser/read.pdb respectively
or a pdb ID so that the function can download the data.}

\item{cutoff}{A numeric to set the maximum distance for atoms to be
returned.}

\item{select}{A string that should match "Nuc", "Prot", "DNA" or "RNA", to
be used as reference.}

\item{nchain}{A string with the nucleic acid chain to get data about.
If NULL, all of them are selected (according with select argument).}

\item{pchain}{A string with the protein chain to get data about.
If NULL, all of them are selected.}

\item{hydrogens}{A logical to use the hydrogens in the structure or remove
them.}

\item{byres}{A logical to indicate if the output should be referred to the
residues rather than atoms.}

\item{verbose}{A logical to print details of the process.}

\item{...}{Arguments to selectModel and/or alt records.}
}
\value{
A data.frame with the atomic distances in the interacting site.
}
\description{
For pdb structures with protein-nucleic acid complexes, the function finds
the atoms in the interacting site. It allows the user to set as reference
the nucleic acid, the protein, or particular desired chains.
}
\examples{
    pdb <- cifParser("1b3t") # A protein-DNA complex
    data <- findBindingSite(pdb, select="DNA", byres=TRUE)

}
\author{
Diego Gallego
}
