% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{findHDR}
\alias{findHDR}
\title{Find 2D High Density Regions and return list of nucleotides}
\usage{
findHDR(
  ntID = NULL,
  ntinfo,
  x = "eta",
  y = "theta",
  SD_DENS = 1,
  bandwidths = c(40, 40),
  dens = NULL,
  lims = c(0, 360, 0, 360)
)
}
\arguments{
\item{ntID}{an object of class integer with the desired nucleotides of
analysis. If NULL all the nucleotides in the data.frame will be used.}

\item{ntinfo}{a data.frame with the input data. It should contain three
columns (additional columns will be ignored). One of them should be
"ntID" and the other two are optional and can be specified using the
parameters "x" and "y".}

\item{x}{name of the column that will be used as "x" axis.}

\item{y}{name of the column that will be used as "y" axis.}

\item{SD_DENS}{height above the mean to be used to select the nucleotides}

\item{bandwidths}{object to be passed to the "kde2d" function (only used if
"dens" is NULL).}

\item{dens}{optional object containing the output of "kde2d" or equivalent.}

\item{lims}{The limits of the rectangle covered by the grid as c(xl, xu,
yl, yu).}
}
\value{
a list of vectors containing the nucleotides ID (according with the
ntID column) that clusterize in different regions of the 2D diagram.
}
\description{
Given a data.frame ("ntinfo") the function computes a 2D Kernel Density
Estimation between the desired fields and returns a list of
nucleotides (according with the ntID column) that
are found in High Density Regions of the 2D map.
}
\author{
Diego Gallego
}
