% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{getID}
\alias{getID}
\title{Extract the unique identifier for each of the nucleotides}
\usage{
getID(pdbID = NULL, ntinfo = NULL)
}
\arguments{
\item{pdbID}{A list/vector containing the desired PDB IDs or a list of pdb
objects as provided by "read.pdb", "read.cif", "cifParser" ...}

\item{ntinfo}{A data.frame containing the information retreived from the
PipeNucData function from the veriNa3d package containing unique
parameters for each nucleotide}
}
\value{
A string with the unique identifier for each nucleotide.
}
\description{
Generate the ID for every nucleotide according with dssr format:
model:chain.residresno^insert.
}
\examples{
    pdblist <- list("1bau", "2rn1")
    ntinfo <- .extract_ntinfo(pdbID=pdblist)
    id_dssr <- id_dssr_extract(ntinfo)
    ntinfo$id_dssr <- id_dssr #Both id_dssr and ntinfo have the same length

}
\author{
Eric Matamoros & Diego Gallego
}
