% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRNAList.R
\name{getRNAList}
\alias{getRNAList}
\title{Download Representative List of RNA structures}
\usage{
getRNAList(release = "current", threshold = "all", as.df = FALSE)
}
\arguments{
\item{release}{A number indicating the list of interest.}

\item{threshold}{A string that matches one of the lists in the BGSU RNA
site ("1.5A", "2.0A", "2.5A", "3.0A", "3.5A", "4.0A", "20.0A", "all").
Note that "all" returns structures solved by any technique.}

\item{as.df}{A logical to return the output as a data.frame}
}
\value{
A data frame with the list of Equivalence Classes and the
Representant and Members of each Equivalence Class. Note that the
output is formated according with Leontis&Zirbel nomenclature
(AAAA|M|C), where "AAAA" is the PDB accession code, "M" is the model
and "C" is the Chain to be used.
}
\description{
According with Leontis & Zirbel (references below), the PDB contains
structures that are redundant. Their server provides weekly releases
of Representative Sets of RNA structures (also called non-redundant
lists). This function access their website and returns the desired list.
}
\examples{
    data <- getRNAList(release=3.2, threshold="1.5A")

}
\references{
\preformatted{Official site: http://rna.bgsu.edu/rna3dhub/nrlist/
Publication: Leontis, N.B., and C.L. Zirbel. 2012. “Nonredundant 3D 
Structure Datasets for RNA Knowledge Extraction and Benchmarking.” 
In RNA 3D Structure Analysis and Prediction, edited by N. Leontis and
E. Westhof, 27:281–98. Springer Berlin Heidelberg
}
}
\author{
Diego Gallego
}
