% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasHetAtm.R
\name{hasHetAtm}
\alias{hasHetAtm}
\title{Check if a given PDB contains the ligand/modbase of interest}
\usage{
hasHetAtm(pdbID, hetAtms)
}
\arguments{
\item{pdbID}{A 4-character string.}

\item{hetAtms}{A string with the ligand/modbase ID.}
}
\value{
A logical. TRUE if the given hetAtms is present in the structure.
}
\description{
Given a 4-character string (PDB ID) and a ligand/modbase ID, the function
checks the presence of the ligand/modres in the given PDB ID. To check for
the presence of sodium ions use hetAtms="Na" instead of NA. If you are
interested on the whole list of heterogeneous atoms see
\code{\link[=queryHetAtms]{queryHetAtms()}}.
}
\examples{
hasHetAtm("1s72", "MG") # Check if structure has Magnesium ion

}
\author{
Diego Gallego
}
