% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureElenoDist.R
\name{measureElenoDist}
\alias{measureElenoDist}
\title{Computes distances between the atoms of interest in a mmCIF structure}
\usage{
measureElenoDist(
  pdb,
  model = NULL,
  refeleno,
  eleno,
  n = 1,
  cutoff = c(0, 5),
  verbose = FALSE,
  detailedoutput = TRUE,
  data_of_interest = NULL
)
}
\arguments{
\item{pdb}{A pdb object obtained as from
\code{\link[=cifAsPDB]{cifAsPDB()}}
or read.pdb/read.cif (bio3d functions).}

\item{model}{The model of interest to use in the calculations. The first
model is always the default.}

\item{refeleno}{A vector of eleno (element number) to take as reference.}

\item{eleno}{A vector of eleno to measure the distances.}

\item{n}{An integer indicating how many closests atoms to return. The
default n=1 returns only the closest atom; n=2 would return the two
closest atoms and so on. If NULL, any number of atoms within the
cutoff will be returned.}

\item{cutoff}{A numeric vector indicating the distance range to consider in
angstroms. Atoms further than the cutoff won't be returned.}

\item{verbose}{A logical indicating whether to print details of the process.}

\item{detailedoutput}{A logical indicating whether to include additional
information for each atom (see data_of_interest below). If FALSE, only
the eleno (element number) and distances are returned.}

\item{data_of_interest}{A vector of strings. Only used if detailedoutput is
TRUE. The vector should only contain the strings between the following:
"type", "elety", "alt", "resid", "chain", "resno", "insert", "x", "y",
"z", "o", "b", "entid", "elesy", "charge", "asym_id", "seq_id",
"comp_id", "atom_id", "model".
The selected fields will be returned for both atoms.}
}
\value{
A data.frame with the nearest atom neighbours information. Fields
suffixed with '_A' refer to the atoms used as reference. Fields
suffixed with '_B' refer to the 'contacting'/closest atoms.
}
\description{
Given a pdb object (or a pdb ID), the function computes the distances
between the desired atoms and returns the closest ones. Note that eleno
numbers might be different in the PDB vs mmCIF formats and this may lead to
errors.
}
\examples{
    ## Dowload cif file and save coordinates data
    cif <- cifParser("1enn")
    coordinates <- cifAtom_site(cif)

    ## Find atom numbers for desired entities (e.g. water and DNA)
    water_eleno <- coordinates[coordinates$label_atom_id == "O", "id"]
    dna_eleno <- coordinates[coordinates$label_comp_id \%in\% 
                                        c("DA", "DT", "DG", "DU"), "id"]

    ## Find which DNA atoms are in 5 Angstroms distance from the water
    data <- measureElenoDist(cif, refeleno=water_eleno, eleno=dna_eleno, 
                                n=NULL, cutoff=5)

    ## To see the data
    head(data)

}
\author{
Diego Gallego
}
