% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureNuc.R
\name{measureNuc}
\alias{measureNuc}
\title{Obtain desired nucleotide measurements}
\usage{
measureNuc(
  pdb,
  model = 1,
  chain = "all",
  v_shifted = TRUE,
  b_shifted = TRUE,
  distances = "default",
  angles = "default",
  torsionals = "default",
  pucker = TRUE,
  Dp = TRUE,
  refatm = "C4'",
  force = FALSE
)
}
\arguments{
\item{pdb}{A pdb object as obtained from cifAsPDB or read.cif/read.pdb
(bio3d package).}

\item{model}{A string with the desired model number.}

\item{chain}{A string with the desired chain id.}

\item{v_shifted}{A logical. If TRUE, puckering angles (nu0 to nu4) are
returned in the range 0 to 360 degrees. Otherwise, -180 to +180.}

\item{b_shifted}{A logical. If TRUE, backbone angles, chi and kappa are
returned in the range 0 to 360 degrees. Otherwise, -180 to +180.}

\item{distances}{A data.frame indicating all the intra and inter-nucleotide
atomic distances of interest. See details section. A default option is
preconfigured to simplify the use of the function and can be seen
typing 'veriNA3d::.distances'.}

\item{angles}{A data.frame indicating all the intra and inter-nucleotide
angles of interest. See details section. A default option is
preconfigured to simplify the use of the function and can be seen
typing 'veriNA3d::.angles'.}

\item{torsionals}{A data.frame indicating all the intra and inter-
nucleotide torsional angles of interest. See details section. A default
option is preconfigured to simplify the use of the function and can be
seen typing 'veriNA3d::.torsionals'.}

\item{pucker}{A logical indicating whether to compute the puckering.}

\item{Dp}{A logical indicating whether to compute the Dp distance.}

\item{refatm}{A string with the atom to use to identify the nucelotides.
Important to analyse models with just (in example) phosphate atoms,
in which refatm should be set to "P" (it was thought when analysing
the structure with PDB code: 1Y1Y).}

\item{force}{A logical to force the analysis. Useful when the function
does not recognise a nucleic acid in the structure (e.g. because all
bases are non-canonical: 1PBL, 1XV6, 1DV4, etc).}
}
\value{
A data.frame with the measurements for every nucleotide.
}
\description{
From a nucleic acid structure (pdb object), it computes the desired
atomic distances, angles, dihedral angles, puckering conformation and Dp
distance (See definition of Dp in MolProbity paper by Chen et al.
2010).
}
\details{
The format of 'distances', 'angles' and 'torsionals' is:
First column should indicate the first atom, second column second
atom (and so on in the case of angles and torsional angles). An extra
last column is optional and should contain the names to identify each
measurement in the output. Plane atom names are interpreted as intra-
nucleotide measurments. For inter-nucleotide measurments use the prefix
"pre_" or "post_" before the atom name. In example, to compute all
inter-phosphate distances, use as argument: \cr
distances=data.frame(atomA=c("P"), atomB=c("post_P"),
labels=c("interphosphate"),
stringsAsFactors=FALSE).\cr
}
\examples{
    distances <- data.frame(atomA=c("P"), atomB=c("post_P"), 
                            labels=c("interphosphate"), 
                            stringsAsFactors=FALSE)
    measureNuc(cifAsPDB("1bna"), distances=distances, angles=NULL, 
                    torsionals=NULL, Dp=NULL)

}
\author{
Diego Gallego
}
