% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeNucData.R
\name{pipeNucData}
\alias{pipeNucData}
\title{Obtain nucleotide details from a data set of RNA structures}
\usage{
pipeNucData(
  pdbID,
  model = NULL,
  chain = NULL,
  range = c(3, 1e+05),
  path = ".veriNA3d_mmCIF_files",
  extension = ".cif.gz",
  cores = 1,
  progressbar = TRUE,
  ...
)
}
\arguments{
\item{pdbID}{A list/vector containing the desired PDB IDs or a list of pdb
objects as provided by "read.pdb", "read.cif", "cifParser" ...}

\item{model}{A vector with same length of pdbID containing the
desired model for each pdbID. If all models are desired, use "all".
If no models are specified, the first one will be used for each pdbID.}

\item{chain}{A vector with same length of pdbID containing the
desired chain for each pdbID. If no chain is specified, all chains will
be analysed by default. Non-nucleic acid chains will be ignored.}

\item{range}{A numeric vector with two values to indicate the desired
length range for the Nucleic Acid chains. If a chain falls outside the
range, it is not analysed.}

\item{path}{Directory in which the PDB/CIF files can be found (if NULL, the
function will download them). If you provide a "path", make sure the
file names are the PDB IDs followed by ".cif" or "pdb". The function
will find them using the strings in pdbID, so make sure you use the
same case.}

\item{extension}{A string matching the files extension (e.g. ".pdb",
".cif", "pdb.gz", "cif.gz").
Only necessary if the PDB files are to be read from disk and a path is
provided.}

\item{cores}{Number of CPU cores to be used.}

\item{progressbar}{A logical to print in screen a progress bar.}

\item{...}{Arguments to be passed to \code{\link[=measureNuc]{measureNuc()}}.}
}
\value{
A data.frame with data about every nucleotide in the input set.
}
\description{
Pipeline to generate a data.frame with the desired info for a list of PDB.
Nucleotides are labeled with a unique identifier (column ntID).
}
\examples{
    ## This is a toy example, see vignettes for real-world usages.
    pdblist <- list("1bau", "2rn1")
    model <- list("1", "2")
    chain <- list("all", "all")
    ntinfo <- pipeNucData(pdbID=pdblist, model=model, chain=chain)

}
\author{
Diego Gallego
}
