% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeProtNucData.R
\name{pipeProtNucData}
\alias{pipeProtNucData}
\title{Obtain nucleotide-protein interactions from a data set of structures}
\usage{
pipeProtNucData(
  pdbID,
  model = NULL,
  chain = NULL,
  ntinfo = NULL,
  path = NULL,
  extension = NULL,
  cores = 1,
  progressbar = TRUE,
  cutoff = 15,
  ...
)
}
\arguments{
\item{pdbID}{A list/vector containing the desired PDB IDs or a list of pdb
objects as provided by "read.pdb", "read.cif", "cifParser" ...}

\item{model}{A vector with same length of pdbID containing the
desired model for each pdbID. If all models are desired, use "all".
If no models are specified, the first one will be used for each pdbID.}

\item{chain}{A vector with same length of pdbID containing the
desired chain for each pdbID. If no chain is specified, all chains will
be analysed by default. Non-nucleic acid chains will be ignored.}

\item{ntinfo}{Optional. A data.frame obtained from
\code{\link[=pipeNucData]{pipeNucData()}}
for the same dataset (or bigger), but not smaller.}

\item{path}{Directory in which the PDB/CIF files can be found (if NULL, the
function will download them). If you provide a "path", make sure the
file names are the PDB IDs followed by ".cif" or "pdb". The function
will find them using the strings in pdbID, so make sure you use the
same case.}

\item{extension}{A string matching the files extension (e.g. ".pdb",
".cif", "pdb.gz", "cif.gz").
Only necessary if the PDB files are to be read from disk and a path is
provided.}

\item{cores}{Number of CPU cores to be used.}

\item{progressbar}{A logical to print in screen a progress bar.}

\item{cutoff}{A numeric with the maximum distance to return. To be passed
to \code{\link[=findBindingSite]{findBindingSite()}}.}

\item{...}{Additional arguments to be passed to
\code{\link[=findBindingSite]{findBindingSite()}}.}
}
\value{
A data.frame with data about the atomic distances in the
interacting sites of every structure in the input set.
}
\description{
Pipeline to generate a data.frame with the data about the closests
nucleotides to the protein for a list of PDB.
The data can be related to unique nucleotide
indentifiers (ntID) by providing the output of the independent
pipeline \code{\link[=pipeNucData]{pipeNucData()}}.
}
\examples{
    ## This is a toy example, see vignettes for more usages.
    pdblist <- list("1nyb", "2ms1")
    aantinfo <- pipeProtNucData(pdbID=pdblist)

}
\author{
Diego Gallego
}
