% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot2D}
\alias{plot2D}
\alias{plot3Ddens}
\title{Make 2D data plot or 3D view of the density}
\usage{
plot2D(
  ntinfo,
  ntID = NULL,
  dens = NULL,
  bandwidths = NULL,
  x = "eta",
  y = "theta",
  drawcontour = TRUE,
  sd_over_mean_contours = c(1, 2, 4),
  etatheta = FALSE,
  points = NULL,
  colpoints = "red",
  file = NULL,
  width = 15,
  height = 15,
  bg = "white",
  units = "cm",
  res = 200,
  xlab = "default",
  ylab = "default"
)

plot3Ddens(
  ntinfo,
  ntID = NULL,
  dens = NULL,
  bandwidths = NULL,
  x = "eta",
  y = "theta",
  defaultview = NULL,
  thetaplot,
  phiplot,
  cleanerview = FALSE,
  file = NULL,
  width = 15,
  height = 15,
  bg = "white",
  units = "cm",
  res = 600,
  xlab = "default",
  ylab = "default",
  zlab = "default"
)
}
\arguments{
\item{ntinfo}{A data.frame with the input data. It should contain the
columns with the desired data and a column labeled ntID.}

\item{ntID}{A vector of integers with the desired nucleotides of
analysis. If NULL all the nucleotides in the data.frame will be used.}

\item{dens}{The output of a kernel density estimation (e.g. kde2d function)
over the data of interest. If it is NULL and drawcontour=TRUE, it will
be computed on the fly.}

\item{bandwidths}{In case dens=NULL and drawcontour=TRUE, it will be passed
to kde2d().}

\item{x}{A string with the parameter to be placed in the x axis.}

\item{y}{A string with the parameter to be placed in the y axis.}

\item{drawcontour}{A logical to highlight the most populated regions of the
plot.}

\item{sd_over_mean_contours}{A numeric vector with the standard deviations
over the mean to plot the contours (in case drawcontour=TRUE).}

\item{etatheta}{A logical to return a eta-theta plot.}

\item{points}{A vector of integers. It should contain the ntID of the
nucleotides to print in a different color.}

\item{colpoints}{A string with the desired color if points is not NULL.}

\item{file}{A string with the name of the output file. If NULL, the
plot will be printed to screen.}

\item{width}{The width of the plot (passed to the png() function).}

\item{height}{The height of the plot (passed to the png() function).}

\item{bg}{The background color of the plot (passed to the png() function).}

\item{units}{The unit to measure height and width (passed to the png()
function).}

\item{res}{Resolution (passed to the png() function).}

\item{xlab}{String for X axis label}

\item{ylab}{String for Y axis label}

\item{defaultview}{A string to set different default options. Chose
between '2Dupview', 'leftview' and 'rightview'.}

\item{thetaplot}{Argument to be pased to \code{persp3D()} to set the view
perspective.}

\item{phiplot}{Argument to be pased to \code{persp3D()} to est the view
perspective.}

\item{cleanerview}{A logical to remove the lower density region to get a
cleaner plot.}

\item{zlab}{String for Z axis label}
}
\value{
A plot in screen, which can be directly saved  to a ".png" file.
* {plot2D} A scatter plot.
* {plot3Ddens} A density map of the data in 3D.
}
\description{
Function to make a plot highlighting the most populated regions.
}
\examples{
    ntinfo <- pipeNucData("1bau")
    C3endo_ntID <- cleanByPucker(ntinfo, pucker="C3'endo")
    plot2D(ntinfo=ntinfo, ntID=C3endo_ntID)

}
\author{
Diego Gallego
}
