% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAPI.R
\name{queryAPI}
\alias{queryAPI}
\title{Launch queries to the MMB or EBI APIs}
\usage{
queryAPI(
  ID,
  info = NULL,
  API = "default",
  string1 = NULL,
  string2 = NULL,
  reuse = TRUE,
  envir = parent.frame(n = 2),
  verbose = FALSE,
  noerrors = FALSE,
  https = FALSE
)
}
\arguments{
\item{ID}{A 4 character string that matches a structure in the Protein
Data Bank, or a 3 character string matching a compound.}

\item{info}{A string with the desired query name.}

\item{API}{A string that matches "ebi" or "mmb".}

\item{string1}{A string to configure the query. See example below.}

\item{string2}{A string to configure the query. See example below.}

\item{reuse}{A logical. Set to TRUE if the same query is going to be send
repeatedly, so that the result is saved in RAM (ir provides faster
user access and avoids unnecessary work in the servers).}

\item{envir}{Environment to save&retrieve the data if reuse is TRUE.}

\item{verbose}{A logical. TRUE to print details of the process.}

\item{noerrors}{A logical. TRUE to suppress errors.}

\item{https}{A logical. TRUE to use https instead of http.}
}
\value{
A vector or data.frame with the desired data.
}
\description{
Given a 4-character string (PDB ID) and the desired "info", it sends a
query to the desired API and returns the output. This is an intermediate
wrapper called by most of the queryFunctions (for documentation see
?queryFunctions).
}
\examples{
    ## Imagine you want to programmatically access the EBI API contents
    ## through "http://www.ebi.ac.uk/pdbe/api/topology/entry/1s72/chain/H".
    ## 'queryAPI' understands it with four intructions:
    ## 'API="ebi"' stands for the root of the website name ("http.../api/").
    ## 'string1' is the string from the root to the pdb ID.
    ## 'ID' is just the PDB code.
    ## 'string2' is the string after the pdb ID.
    ## Thus, the call would be:
    #data <- queryAPI(ID="1s72", API="ebi", 
    #                    string1="topology/entry/", string2="chain/H/")

}
\author{
Diego Gallego
}
