% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryEntryList.R
\name{queryEntryList}
\alias{queryEntryList}
\title{Downloads the list of ID of ALL current PDB entries}
\usage{
queryEntryList(justIDs = TRUE)
}
\arguments{
\item{justIDs}{A logical to return only pdbIDs without other information.}
}
\value{
A vector with all the PDB ID entries (updated weekly).
}
\description{
Function to get the list of ALL PDB IDs in the Protein Data Bank at the
moment.
}
\examples{
# pdblist <- queryEntryList()

}
\author{
Diego Gallego
}
