% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryObsoleteList.R
\name{queryObsoleteList}
\alias{queryObsoleteList}
\title{Downloads the list of ID of ALL current PDB entries}
\usage{
queryObsoleteList()
}
\value{
A vector with all the PDB ID entries (updated weekly).
}
\description{
Function to get the list of Obsolete PDB IDs in the Protein Data Bank at
the moment.
}
\examples{
# obsolete <- queryObsoleteList()

}
\author{
Diego Gallego
}
