% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-refframes.R
\docType{data}
\name{refframes}
\alias{refframes}
\title{PDB objects for all canonical bases with reference frame}
\format{An object of class \code{list} of length 12.}
\usage{
data("refframes")
}
\value{
Multiple objects with pdb format
}
\description{
PDB objects for all canonical bases with reference frame
}
\keyword{datasets}
