% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/represAsDataFrame.R
\name{represAsDataFrame}
\alias{represAsDataFrame}
\title{Coerce Representative list to a data.frame}
\usage{
represAsDataFrame(nrlist)
}
\arguments{
\item{nrlist}{The output of
\code{\link[=getRNAList]{getRNAList()}} or
\code{\link[=getAltRepres]{getAltRepres()}}.}
}
\value{
A data frame with the data of the representative structures.
}
\description{
Takes the output of getRNAList or getAltRepres, which
represent molecules with the format "XXXX|M|C+XXXX|M|C" (XXXX: PDB ID;
M: Model; C: Chain) and returns a data.frame with a more friendly
structure:
\itemize{
\item Col 1: Equivalence Class.
\item Col 2: PDB ID.
\item Col 3: Model.
\item Col 4: Chain.
}
Columns 2 to 4 can be the direct input of
\code{\link[=pipeNucData]{pipeNucData()}}.
}
\examples{
 data <- getRNAList(release=3.2, threshold="1.5A")
 reps <- represAsDataFrame(nrlist=data)

}
\author{
Diego Gallego
}
