% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-CIF.R
\name{selectModel}
\alias{selectModel}
\alias{selectModel,CIF-method}
\alias{selectModel,ANY-method}
\title{Selects a desired model from a CIF/pdb structure}
\usage{
selectModel(cif, pdb, model, verbose = FALSE)

\S4method{selectModel}{CIF}(cif, model, verbose = FALSE)

\S4method{selectModel}{ANY}(pdb, model, verbose = FALSE)
}
\arguments{
\item{cif}{A CIF object as otained from cifParser.}

\item{pdb}{A pdb object with multiple models (obtained from cifAsPDB
or read.cif/read.pdb from bio3d package).}

\item{model}{A string with the desired model number.}

\item{verbose}{A logical to print messages on screen.}
}
\value{
A CIF/pdb object with the desired model coordinates.
}
\description{
Given a object obtained from cifParser/cifAsPDB or read.cif/read.pdb
functions (from bio3d specifying "multi = TRUE"), the function returns an
object of the same type (CIF or pdb) with the desired model.
Since some structures deposited in the PDB contain models with different
number of atoms, the pdb objects in R require a special treatment.
The cifAsPDB and selectModel functions can cope with these structures
(e.g. 1JTW).
}
\examples{
    cif <- cifParser("1qfq")
    model3 <- selectModel(cif=cif, model=3)

}
\author{
Diego Gallego
}
