% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veriNA3d-package.R
\docType{package}
\name{veriNA3d}
\alias{veriNA3d}
\title{veriNA3d: Structural bioinformatics package for data mining of the PDB.}
\description{
VeriNA3d is an R package for the analysis of Nucleic Acid structural data.
The software was developed on top of bio3d (Grant et al, 2006) with a higher
level of abstraction. In addition of single-structure analyses, veriNA3d also
implements pipelines to handle whole datasets of mmCIF/PDB structures. As far
as we know, no similar software has been previously distributed, thus it aims
to fill a gap in the data mining pipelines of PDB structural data analyses.
}
\author{
Diego Gallego
\href{mailto:diego.gallego@irbbarcelona.org}{diego.gallego@irbbarcelona.org}
}
